+++ Sources/Controls/XYPad.swift
@@ -1,10 +1,10 @@
 import SwiftUI
 
 public struct XYPad: View {
-    @Binding var x: Double
-    @Binding var y: Double
+    @Binding var x: Float
+    @Binding var y: Float
 
-    public init(x: Binding<Double>, y: Binding<Double>) {
+    public init(x: Binding<Float>, y: Binding<Float>) {
         _x = x
         _y = y
     }
@@ -15,8 +15,8 @@ public struct XYPad: View {
                 RoundedRectangle(cornerRadius: geo.size.width / 20).foregroundColor(.gray)
                 Circle().foregroundColor(.red)
                     .frame(width: geo.size.width / 10, height: geo.size.height / 10)
-                    .offset(x: x * (geo.size.width - geo.size.width / 10),
-                            y: -y * (geo.size.height - geo.size.height / 10))
+                    .offset(x: CGFloat(x) * (geo.size.width - geo.size.width / 10),
+                            y: CGFloat(-y) * (geo.size.height - geo.size.height / 10))
             }
         }
     }
